Rename march_api to march_main. Split the code into a src and api folder
