--- march_ws/src/march_api/launch/api.launch
@@ -1,3 +1,0 @@
-<launch>
-    <node name="api_node" pkg="march_api" type="api_node" output="screen"/>
-</launch>
