+++ march_ws/src/march_main/api/main.cpp
@@ -2,8 +2,8 @@
 
 #include "std_msgs/Bool.h"
 #include "ros/ros.h"
-#include "march_api/Trigger.h"
-#include "../common/communication/TopicNames.h"
+#include "march_main/Trigger.h"
+#include "../src/common/communication/TopicNames.h"
 #include "LaunchAPI.cpp"
 
 int main(int argc, char** argv)
