+++ march_ws/src/march_state_machine/CMakeLists.txt
@@ -8,7 +8,7 @@ add_compile_options(-std=c++11)
 ## if COMPONENTS list like find_package(catkin REQUIRED COMPONENTS xyz)
 ## is used, also find other catkin packages
 find_package(catkin REQUIRED COMPONENTS
-        march_api
+        march_main
         rospy
         smach
         smach_ros
