+++ march_ws/src/march_state_machine/launch/smach.launch
@@ -1,5 +1,5 @@
 <launch>
-    <include file="$(find march_api)/launch/api.launch"/>
+    <include file="$(find march_main)/launch/api.launch"/>
 
     <node name="State_Machine" pkg="march_state_machine" type="StateMachine.py" output="screen"/>
     <node name="State_Machine_viewer" pkg="smach_viewer" type="smach_viewer.py" output="screen"/>
