+++ march_ws/src/march_state_machine/src/launch_sm/UrdfState.py
@@ -1,7 +1,8 @@
 import rospy
 import smach
-from march_api.srv import Trigger
+from march_main.srv import Trigger
 
+import time
 
 class UrdfState(smach.State):
     def __init__(self):
@@ -9,7 +10,9 @@ class UrdfState(smach.State):
 
     def execute(self, userdata):
         rospy.loginfo('Checking config')
+        time.sleep(5)
         checkUrdf = rospy.ServiceProxy('march/urdf_validation', Trigger)
+
         result = checkUrdf()
         rospy.loginfo(result)
         if result.success:
